Add TransferTest and CouponTest
